// sunrei-server/src/dto/models/SunreiDTO.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Sunrei API
 * API documentation for Sunrei project
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { SunreiSpotDTO } from './SunreiSpotDTO';
import {
    SunreiSpotDTOFromJSON,
    SunreiSpotDTOFromJSONTyped,
    SunreiSpotDTOToJSON,
    SunreiSpotDTOToJSONTyped,
} from './SunreiSpotDTO';
import type { TagDTO } from './TagDTO';
import {
    TagDTOFromJSON,
    TagDTOFromJSONTyped,
    TagDTOToJSON,
    TagDTOToJSONTyped,
} from './TagDTO';
import type { ImageDTO } from './ImageDTO';
import {
    ImageDTOFromJSON,
    ImageDTOFromJSONTyped,
    ImageDTOToJSON,
    ImageDTOToJSONTyped,
} from './ImageDTO';

/**
 * 
 * @export
 * @interface SunreiDTO
 */
export interface SunreiDTO {
    /**
     * 
     * @type {string}
     * @memberof SunreiDTO
     */
    id?: string;
    /**
     * Title of the Sunrei
     * @type {string}
     * @memberof SunreiDTO
     */
    title?: string;
    /**
     * Detailed description of the Sunrei
     * @type {string}
     * @memberof SunreiDTO
     */
    description?: string;
    /**
     * External link for this Sunrei (program URL or YouTube channel)
     * @type {string}
     * @memberof SunreiDTO
     */
    link?: string | null;
    /**
     * Array of images for this Sunrei
     * @type {Array<ImageDTO>}
     * @memberof SunreiDTO
     */
    images?: Array<ImageDTO>;
    /**
     * Array of Sunrei spots associated with this Sunrei
     * @type {Array<SunreiSpotDTO>}
     * @memberof SunreiDTO
     */
    spots?: Array<SunreiSpotDTO>;
    /**
     * Array of tags associated with this Sunrei
     * @type {Array<TagDTO>}
     * @memberof SunreiDTO
     */
    tags?: Array<TagDTO>;
}

/**
 * Check if a given object implements the SunreiDTO interface.
 */
export function instanceOfSunreiDTO(value: object): value is SunreiDTO {
    return true;
}

export function SunreiDTOFromJSON(json: any): SunreiDTO {
    return SunreiDTOFromJSONTyped(json, false);
}

export function SunreiDTOFromJSONTyped(json: any, ignoreDiscriminator: boolean): SunreiDTO {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'] == null ? undefined : json['id'],
        'title': json['title'] == null ? undefined : json['title'],
        'description': json['description'] == null ? undefined : json['description'],
        'link': json['link'] == null ? undefined : json['link'],
        'images': json['images'] == null ? undefined : ((json['images'] as Array<any>).map(ImageDTOFromJSON)),
        'spots': json['spots'] == null ? undefined : ((json['spots'] as Array<any>).map(SunreiSpotDTOFromJSON)),
        'tags': json['tags'] == null ? undefined : ((json['tags'] as Array<any>).map(TagDTOFromJSON)),
    };
}

export function SunreiDTOToJSON(json: any): SunreiDTO {
    return SunreiDTOToJSONTyped(json, false);
}

export function SunreiDTOToJSONTyped(value?: SunreiDTO | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'title': value['title'],
        'description': value['description'],
        'link': value['link'],
        'images': value['images'] == null ? undefined : ((value['images'] as Array<any>).map(ImageDTOToJSON)),
        'spots': value['spots'] == null ? undefined : ((value['spots'] as Array<any>).map(SunreiSpotDTOToJSON)),
        'tags': value['tags'] == null ? undefined : ((value['tags'] as Array<any>).map(TagDTOToJSON)),
    };
}